Refactor: Adapter repackage
