+++ src/main/java/shop/s5g/front/adapter/delivery/DeliveryAdapter.java
@@ -1,4 +1,4 @@
-package shop.s5g.front.adapter;
+package shop.s5g.front.adapter.delivery;
 
 import org.springframework.cloud.openfeign.FeignClient;
 import org.springframework.web.bind.annotation.PutMapping;
