+++ src/main/java/shop/s5g/front/adapter/point/PointPolicyAdapter.java
@@ -1,4 +1,4 @@
-package shop.s5g.front.adapter;
+package shop.s5g.front.adapter.point;
 
 import java.util.List;
 import org.springframework.cloud.openfeign.FeignClient;
