+++ src/main/java/shop/s5g/front/service/delivery/impl/DeliveryFeeServiceImpl.java
@@ -5,7 +5,7 @@ import java.util.concurrent.CompletableFuture;
 import lombok.RequiredArgsConstructor;
 import org.springframework.scheduling.annotation.Async;
 import org.springframework.stereotype.Service;
-import shop.s5g.front.adapter.DeliveryFeeAdapter;
+import shop.s5g.front.adapter.delivery.DeliveryFeeAdapter;
 import shop.s5g.front.dto.delivery.DeliveryFeeCreateRequestDto;
 import shop.s5g.front.dto.delivery.DeliveryFeeResponseDto;
 import shop.s5g.front.dto.delivery.DeliveryFeeUpdateRequestDto;
