+++ src/main/java/shop/s5g/front/service/delivery/impl/DeliveryServiceImpl.java
@@ -2,7 +2,7 @@ package shop.s5g.front.service.delivery.impl;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
-import shop.s5g.front.adapter.DeliveryAdapter;
+import shop.s5g.front.adapter.delivery.DeliveryAdapter;
 import shop.s5g.front.dto.delivery.DeliveryResponseDto;
 import shop.s5g.front.dto.delivery.DeliveryUpdateRequestDto;
 import shop.s5g.front.service.delivery.DeliveryService;
